requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bitarray==2.7.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
cytoolz==0.12.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
ghp-import==2.1.0
hexbytes==0.3.0
idna==3.4
importlib-metadata==4.13.0
ipfshttpclient==0.8.0a2
Jinja2==3.1.2
jsonschema==4.17.3
linkify-it-py==2.0.0
lru-dict==1.1.8
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdit-py-plugins==0.3.4
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-exclude==1.0.2
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==1.0.0
netaddr==0.8.0
packaging==23.0
parsimonious==0.8.1
pathspec==0.11.0
platformdirs==3.0.0
protobuf==3.19.5
py-solc-x==1.1.1
pycparser==2.21
pycryptodome==3.17
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.28.2
rich==13.3.1
rlp==2.0.1
semantic-version==2.10.0
six==1.16.0
textual==0.12.1
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.14
varint==1.0.2
watchdog==2.3.0
web3==5.31.3
websockets==9.1
yarl==1.8.2
zipp==3.15.0
